Extract getTopPlayers from Finale.
